refactor(voyage): les frais d'autoload sortent d'app.js
`frais.ts` porte les 121 lignes du bloc : le store des relevés (`loadAutoloadK`, `saveAutoloadK`),
la résolution du coefficient (`globalK`, `kFor`), le contexte de calcul (`feeEnd`, `feeCtx`,
`feeResolver` — passée dix fois aux fonctions du moteur), et les textes qui expliquent le chiffre.
`app.js` : 3 346 → 3 218 lignes, et quatre imports de `logic.ts` devenus morts partent avec.

Le CALCUL était déjà pur et dans `logic.ts` ; ce bloc n'apportait que ce que `logic.ts` ne peut pas
deviner — quel terminal porte quel nom, et combien CETTE station facture. Il ne peignait rien : il
n'avait rien à faire dans la coquille.

`attr` EST SUPPRIMÉ. `feeCell` rendait `{ attr, mark, text }` où `attr` était une chaîne HTML
` title="…"` échappée à la main. Plus aucune vue ne la lit — elles prennent le TEXTE et laissent
React poser l'attribut, comme le dit déjà le commentaire de `vues/boucles.tsx`. Le champ part du
type `CelluleFrais` des trois vues qui le déclaraient, et `esc` avec lui : ce module n'échappe plus
rien, parce qu'il n'écrit plus de HTML.

UNE ERREUR EN CHEMIN, ET CE QU'ELLE APPREND SUR L'OUTILLAGE. `globalK` et `kFmt` restaient utilisés
par `app.js` et je ne les avais pas exportés. Un identifiant libre dans un module ES est une
`ReferenceError` À L'EXÉCUTION : `vite build` ne la voit pas — il ne résout pas les identifiants
libres — et `tsc` ne regarde pas `app.js`. **Seule la suite e2e l'a dite**, par deux échecs (le
relevé d'autoload de #24, et les quatre réglages du Plan de vol).

Cela complète la leçon de #135 : le build est le PARSEUR d'`app.js`, mais il n'en est pas le
vérificateur. Après toute sortie de bloc, relever les noms du bloc encore référencés ailleurs — un
`grep` suffit, et il est plus rapide que la suite.

Contribue à #96.
